chore: improve codebase
