+++ components/Avatar/Avatar.common.ts
@@ -1,0 +1,8 @@
+import { User } from '@prisma/client';
+
+export const getImageUrl = (user: User) => {
+  return (
+    user?.image ??
+    `https://avatars.dicebear.com/api/initials/${user?.name}.png?backgroundColorLevel=800&fontSize=40`
+  );
+};
