+++ components/Button/Button.interface.tsx
@@ -1,0 +1,8 @@
+export type ButtonProps = {
+  type?: 'button' | 'submit';
+  onClick: () => void;
+  isDisabled: boolean;
+  haveAnimation?: boolean;
+  className?: string;
+  children: React.ReactNode;
+};
