+++ components/ErrorFallback/ErrorFallback.interface.ts
@@ -1,0 +1,4 @@
+export interface ErrorFallbackProps {
+  error: Error;
+  resetErrorBoundary: () => void;
+}
