+++ components/Form/Input/Input.tsx
@@ -1,16 +1,6 @@
 import { Typography } from '@/components/Typography';
-import React, { InputHTMLAttributes, LegacyRef } from 'react';
-import { FieldError } from 'react-hook-form';
-
-interface InputProps extends InputHTMLAttributes<HTMLInputElement> {
-  name: string;
-  placeholder: string | undefined;
-  error: FieldError | undefined;
-  required?: boolean;
-  hintMessage?: string;
-  value: string | undefined;
-  label: string;
-}
+import React, { LegacyRef } from 'react';
+import { InputProps } from './Input.interface';
 
 function MyInput(
   {
