--- components/Form/RichTextEditor/index.ts
@@ -1,1 +1,0 @@
-export * from './RichTextEditor';
