+++ components/Projects/Project.tsx
@@ -20,8 +20,8 @@ export const Project = ({
 
   return (
     <motion.li
-      initial={{ opacity: 0 }}
-      animate={{ opacity: 1 }}
+      initial={{ opacity: 0, scale: 0.9 }}
+      animate={{ opacity: 1, scale: 1 }}
       exit={{ opacity: 0 }}
       transition={{ duration: 1 }}
       className="w-full p-2 shadow-lg md:p-0"
