+++ components/SubmitPageForm/SubmitPageForm.interface.ts
@@ -1,0 +1,8 @@
+export type FormInputs = {
+  tags: string[];
+  projectName: string;
+  repositoryLink: string;
+  projectDescription: string;
+  coverImage: string | null;
+  content: string | null;
+};
