+++ components/index.ts
@@ -1,8 +1,22 @@
 export * from './AuthModal';
 export * from './Avatar';
-export * from './ContributorList';
-export * from './Projects';
-export * from './Search';
-export * from './Layouts';
-export * from './Profile/Project';
 export * from './BackToTopButton';
+export * from './Button';
+export * from './BuyMeACoffee';
+export * from './ContributorList';
+export * from './ErrorFallback';
+export * from './FloatingButtonMenu';
+export * from './Form';
+export * from './LandingPage';
+export * from './Layouts';
+export * from './Markdown';
+export * from './ProductHunt';
+export * from './Profile';
+export * from './Projects';
+export * from './Stats';
+export * from './SubmitPageForm';
+export * from './Tags';
+export * from './ThemeToggler';
+export * from './Toaster';
+export * from './Topbar';
+export * from './Typography';
