+++ public/animations/error.json
