+++ public/animations/hero.json
